Rename sample data helpers to tracing
Rename `helpers.py` to `tracing.py`, to match the metrics helpers.
